Changes for version 3.0:
Add section on automatic constraints to the reference manual
update the chord tolerance section of the reference manual
update reference section on extrude and add revolve and helix.
Add a paragraph on IDF file support.
update the download page to include reference to the VC OMP download.
Mention the snap store for Linux.